Avoid panics from type assertions on nils
If a top-level list of strings contains an empty string, the provider
helper library seems to report a nil instead of an empty string. The
issue does not reproduce for sets of strings, or for lists of strings
that are inside another object.

The empty string doesn't actually make sense, but we don't want to blow
up with a panic. A provider panic causes terraform to report an
extremely scary error message. We'd rather let the JSON go to SFx where
it'll be rejected with a 4xx, whose error output is much more
approachable.