Mostly failing test for colorScaleOptions
This test sometimes passes, but mostly fails. The reason is the
iteration through the ChartColors map to find the palleteIndex - which
is explained nicely in
https://nathanleclaire.com/blog/2014/04/27/a-surprising-feature-of-golang-that-colored-me-impressed/

I did try creating my own *schema.ResourceData struct, but doing so
seemed to require a lot of knowledge of terraform internals and couple
the test to terraform internals. It seemed better and easier to test to
decouple the terraform internals from our logic and split out into two
functions - there may be a more idiomatic way of doing this.